+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## v1.0.3 (2025-06-21)
+
+- src/routes/(default)/login/+page.svelte
+- src/lib/api/UserApi.js => add userLogin
+
 ## v1.0.2 (2025-06-13)
 
 - npm install sweetalert2